projects: ada4250_ardz: restruct to EXAMPLE build
Move ADA4250 IIO app to the EXAMPLE-based flow (common/examples/platform)
mirroring eval-adxrs290-pmdz, and add a project README. Behavior
unchanged.
