+++ doc/sphinx/source/projects/eval/ada4250_ardz.rst
@@ -1,0 +1,1 @@
+.. include:: ../../../../../projects/ada4250_ardz/README.rst
