+++ projects/ada4250_ardz/src/examples/iio_example/example.mk
@@ -1,0 +1,4 @@
+IIOD=y
+
+INCS += $(DRIVERS)/amplifiers/ada4250/iio_ada4250.h
+SRCS += $(DRIVERS)/amplifiers/ada4250/iio_ada4250.c
