+++ projects/ada4250_ardz/src/platform/aducm3029/parameters.h
@@ -1,0 +1,57 @@
+/***************************************************************************//**
+ *   @file   parameters.h
+ *   @brief  Definitions specific to ADuCM3029 platform used by ada4250_ardz
+ *           project.
+ *   @author Antoniu Miclaus (antoniu.miclaus@analog.com)
+********************************************************************************
+ * Copyright 2021(c) Analog Devices, Inc.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of Analog Devices, Inc. nor the names of its
+ *    contributors may be used to endorse or promote products derived from this
+ *    software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
+ * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+ * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*******************************************************************************/
+#ifndef __PARAMETERS_H__
+#define __PARAMETERS_H__
+
+#include <stdint.h>
+#include "platform_init.h"
+#include "aducm3029_spi.h"
+#include "aducm3029_irq.h"
+#include "aducm3029_uart.h"
+
+#define UART_DEVICE_ID	0
+#define UART_IRQ_ID	ADUCM_UART_INT_ID
+#define UART_BAUDRATE	115200
+#define UART_EXTRA	NULL
+#define UART_OPS	&aducm_uart_ops
+
+#define SPI_DEVICE_ID	0
+#define SPI_BAUDRATE	1000000
+#define SPI_CS		1
+#define SPI_OPS		&aducm_spi_ops
+#define SPI_EXTRA	&ada4250_spi_extra_ip
+
+extern struct aducm_spi_init_param ada4250_spi_extra_ip;
+
+#endif /* __PARAMETERS_H__ */
