+++ projects/ada4250_ardz/src/platform/aducm3029/platform_src.mk
@@ -1,0 +1,5 @@
+SRC_DIRS += $(PLATFORM_DRIVERS)
+SRC_DIRS += $(INCLUDE)
+
+IGNORED_FILES += $(PLATFORM_DRIVERS)/aducm3029_uart_stdio.c
+IGNORED_FILES += $(PLATFORM_DRIVERS)/aducm3029_uart_stdio.h
